backend/requirements.txt
fastapi
uvicorn
requests
pandas
numpy
python-dotenv
pydantic
httpx
finvizfinance
yfinance
ta-lib
